remove setting the version from model.CurrentVersion

+++ store/sqlstore/store.go
@@ -948,6 +948,7 @@ func (ss *SqlStore) DropAllTables() {
 			    FROM   pg_class
 			    WHERE  relkind = 'r'  -- only tables
 			    AND    relnamespace = 'public'::regnamespace
+				AND NOT relname = 'db_migrations'
 			   );
 			END
 			$func$;`)
@@ -955,7 +956,9 @@ func (ss *SqlStore) DropAllTables() {
 		tables := []string{}
 		ss.masterX.Select(&tables, `show tables`)
 		for _, t := range tables {
-			ss.masterX.Exec(`TRUNCATE TABLE ` + t)
+			if t != "db_migrations" {
+				ss.masterX.Exec(`TRUNCATE TABLE ` + t)
+			}
 		}
 	}
 }
